refactor(agent): move some types from base.ts to connect.ts
Address, Label, Message, TxHash
