+++ agent/base.ts
@@ -47,20 +47,8 @@ export type Moment = number
 /** A period of time. */
 export type Duration = number
 
-/** A contract's full unique on-chain label. */
-export type Label = string
-
 /** Part of a Label */
 export type Name = string
-
-/** A transaction message that can be sent to a contract. */
-export type Message = string|Record<string, unknown>
-
-/** An address on a chain. */
-export type Address = string
-
-/** A transaction hash, uniquely identifying an executed transaction on a chain. */
-export type TxHash = string
 
 /** Error kinds. */
 class FadromaError extends Error {}
