+++ agent/stub.ts
@@ -1,9 +1,8 @@
 /** Fadroma. Copyright (C) 2023 Hack.bg. License: GNU AGPLv3 or custom.
     You should have received a copy of the GNU Affero General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>. **/
-import type { Address, Message, Label, TxHash } from './base'
 import { assign, Console, Error, base16, sha256, randomBech32 } from './base'
-import type { ChainId } from './connect'
+import type { ChainId, Address, Message, Label, TxHash } from './connect'
 import { Connection, Backend, Batch, Identity } from './connect'
 import type { CodeId, CodeHash } from './deploy'
 import { Compiler, SourceCode, CompiledCode, UploadedCode, ContractInstance } from './deploy'
